don't mask systemd-udev-settle.service
This service was originally masked because after installing the cinnamon desktop, the boot process would often hang on this service
However, this service was started due to multipathd. So the solution for me was to simply remove the device-mapper-multipath package.

Jun 05 03:10:58 fedora udevadm[568]: systemd-udev-settle.service is deprecated. Please fix multipathd.service not to pull it in.

Also, if the systemd-udev-settle.service is masked, it seems to cause issues with /boot and/or /boot/efi not mounting at boot time.